+++ src/jarvis/jarvis/utils/console_utils.py
@@ -1,0 +1,46 @@
+import os
+from subprocess import call
+
+from jarvis._version import __version__
+
+jarvis_logo = "\n" \
+              "      ██╗ █████╗ ██████╗ ██╗   ██╗██╗███████╗\n" \
+              "      ██║██╔══██╗██╔══██╗██║   ██║██║██╔════╝\n" \
+              "      ██║███████║██████╔╝██║   ██║██║███████╗\n" \
+              " ██   ██║██╔══██║██╔══██╗╚██╗ ██╔╝██║╚════██║\n" \
+              " ╚█████╔╝██║  ██║██║  ██║ ╚████╔╝ ██║███████║\n" \
+              "  ╚════╝ ╚═╝  ╚═╝╚═╝  ╚═╝  ╚═══╝  ╚═╝╚══════╝"
+start_text = "" \
+             " -----------------------------------------------\n" \
+             " -  Voice Assistant Platform  " + "v" + __version__ + "-\n" \
+             " -----------------------------------------------\n"
+
+
+class OutputStyler:
+    HEADER = '\033[95m'
+    BLUE = '\033[94m'
+    GREEN = '\033[92m'
+    CYAN = '\033[36m'
+    WARNING = '\033[93m'
+    FAIL = '\033[91m'
+    ENDC = '\033[0m'
+    BOLD = '\033[1m'
+    UNDERLINE = '\033[4m'
+
+
+user_input = OutputStyler.CYAN + '>> ' + OutputStyler.ENDC
+
+
+def clear():
+    """
+    Clear stdout
+    """
+    _ = call('clear' if os.name == 'posix' else 'cls')  # check and make call for specific operating system
+
+
+def stdout_print(text):
+    """
+    Application stdout with format.
+    :param text: string
+    """
+    print(OutputStyler.CYAN + text + OutputStyler.ENDC)
